Upgrade deps - and fix broken integration tests
Due to an increased restriction in hibernate behaviour from a dependency
upgrade, we needed to move some field access inside a transaction, to avoid
a 'no session' hibernate error.
