+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/visit/WithdrawVisitTest.kt
@@ -18,12 +18,10 @@ import uk.gov.justice.digital.hmpps.visitscheduler.dto.OutcomeDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.VisitDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.ApplicationMethodType.WEBSITE
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.OutcomeStatus.REQUESTED_VISIT_WITHDRAWN
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TelemetryVisitEvents
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TelemetryVisitEvents.VISIT_CANCELLED_EVENT
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UnFlagEventReason
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UnFlagEventReason.VISIT_CANCELLED
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType.PUBLIC
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType.STAFF
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitStatus
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitStatus.BOOKED
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitSubStatus
@@ -33,9 +31,7 @@ import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitSubStatus.REQU
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitSubStatus.WITHDRAWN
 import uk.gov.justice.digital.hmpps.visitscheduler.helper.callCancelVisit
 import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
-import uk.gov.justice.digital.hmpps.visitscheduler.service.TelemetryClientService
 import uk.gov.justice.digital.hmpps.visitscheduler.service.VisitNotificationEventService
-import java.time.format.DateTimeFormatter
 
 @DisplayName("Put $VISIT_CANCEL to withdraw a requested visit.")
 class WithdrawVisitTest : IntegrationTestBase() {
@@ -84,7 +80,7 @@ class WithdrawVisitTest : IntegrationTestBase() {
     assertHelper.assertVisitCancellation(visitCancelled, REQUESTED_VISIT_WITHDRAWN, cancelVisitDto.actionedBy, applicationMethodType = WEBSITE, userType = PUBLIC, visitSubStatus = WITHDRAWN)
     assertThat(visitCancelled.visitNotes.size).isEqualTo(0)
 
-    assertTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
+    assertHelper.assertCancelledVisitTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
     assertCancelledDomainEvent(visitCancelled)
     verify(visitNotificationEventServiceSpy, times(1)).deleteVisitAndPairedNotificationEvents(eq(requestedVisit.reference), eq(UnFlagEventReason.REQUESTED_VISIT_WITHDRAWN), eq(null))
   }
@@ -127,7 +123,7 @@ class WithdrawVisitTest : IntegrationTestBase() {
     assertThat(visit1.visitStatus).isEqualTo(visit2.visitStatus)
 
     // just one event thrown
-    assertTelemetryClientEvents(visit1, VISIT_CANCELLED_EVENT)
+    assertHelper.assertCancelledVisitTelemetryClientEvents(visit1, VISIT_CANCELLED_EVENT)
     assertCancelledDomainEvent(visit1)
     verify(visitNotificationEventServiceSpy, times(1)).deleteVisitAndPairedNotificationEvents(eq(requestedVisit.reference), eq(UnFlagEventReason.REQUESTED_VISIT_WITHDRAWN), eq(null))
   }
@@ -168,7 +164,7 @@ class WithdrawVisitTest : IntegrationTestBase() {
     assertHelper.assertVisitCancellation(visitCancelled, REQUESTED_VISIT_WITHDRAWN, cancelVisitDto.actionedBy, applicationMethodType = WEBSITE, userType = PUBLIC, visitSubStatus = WITHDRAWN)
     assertThat(visitCancelled.visitNotes.size).isEqualTo(0)
 
-    assertTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
+    assertHelper.assertCancelledVisitTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
     assertCancelledDomainEvent(visitCancelled)
     verify(visitNotificationEventServiceSpy, times(1)).deleteVisitAndPairedNotificationEvents(eq(requestedVisit.reference), eq(UnFlagEventReason.REQUESTED_VISIT_WITHDRAWN), eq(null))
   }
@@ -208,7 +204,7 @@ class WithdrawVisitTest : IntegrationTestBase() {
     assertHelper.assertVisitCancellation(visitCancelled, REQUESTED_VISIT_WITHDRAWN, cancelVisitDto.actionedBy, applicationMethodType = WEBSITE, userType = PUBLIC, visitSubStatus = VisitSubStatus.CANCELLED)
     assertThat(visitCancelled.visitNotes.size).isEqualTo(0)
 
-    assertTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
+    assertHelper.assertCancelledVisitTelemetryClientEvents(visitCancelled, VISIT_CANCELLED_EVENT)
     assertCancelledDomainEvent(visitCancelled)
     verify(visitNotificationEventServiceSpy, times(1)).deleteVisitAndPairedNotificationEvents(eq(requestedVisit.reference), eq(VISIT_CANCELLED), eq(null))
   }
@@ -260,66 +256,5 @@ class WithdrawVisitTest : IntegrationTestBase() {
       isNull(),
     )
     verify(telemetryClient, times(1)).trackEvent(eq("prison-visit.cancelled-domain-event"), any(), isNull())
-  }
-
-  fun assertTelemetryClientEvents(
-    cancelledVisit: VisitDto,
-    type: TelemetryVisitEvents,
-  ) {
-    val eventAudit = this.eventAuditRepository.findLastEventByBookingReference(cancelledVisit.reference)
-    val visitors = cancelledVisit.visitors.map { visitor -> TelemetryClientService.VisitorDetails(visitor.nomisPersonId.toString(), null) }
-
-    verify(telemetryClient).trackEvent(
-      eq("visit-cancelled"),
-      check {
-        assertThat(it["reference"]).isEqualTo(cancelledVisit.reference)
-        assertThat(it["applicationReference"]).isEqualTo(cancelledVisit.applicationReference)
-        assertThat(it["prisonerId"]).isEqualTo(cancelledVisit.prisonerId)
-        assertThat(it["prisonId"]).isEqualTo(cancelledVisit.prisonCode)
-        assertThat(it["visitStatus"]).isEqualTo(cancelledVisit.visitStatus.name)
-        assertThat(it["visitSubStatus"]).isEqualTo(cancelledVisit.visitSubStatus.name)
-        assertThat(it["visitRestriction"]).isEqualTo(cancelledVisit.visitRestriction.name)
-        assertThat(it["visitStart"]).isEqualTo(cancelledVisit.startTimestamp.format(DateTimeFormatter.ISO_DATE_TIME))
-        assertThat(it["visitEnd"]).isEqualTo(cancelledVisit.endTimestamp.format(DateTimeFormatter.ISO_DATE_TIME))
-        assertThat(it["visitType"]).isEqualTo(cancelledVisit.visitType.name)
-        assertThat(it["visitRoom"]).isEqualTo(cancelledVisit.visitRoom)
-        assertThat(it["hasPhoneNumber"]).isEqualTo(((cancelledVisit.visitContact.telephone != null).toString()))
-        assertThat(it["hasEmail"]).isEqualTo(((cancelledVisit.visitContact.email != null).toString()))
-        assertThat(it["totalVisitors"]).isEqualTo(cancelledVisit.visitors.size.toString())
-        assertThat(it["visitors"]).isEqualTo(objectMapper.writeValueAsString(visitors))
-        eventAudit.actionedBy.userName?.let { value ->
-          assertThat(it["actionedBy"]).isEqualTo(value)
-        }
-        assertThat(it["source"]).isEqualTo(eventAudit.actionedBy.userType.name)
-        assertThat(it["applicationMethodType"]).isEqualTo(eventAudit.applicationMethodType.name)
-        assertThat(it["outcomeStatus"]).isEqualTo(cancelledVisit.outcomeStatus?.name)
-      },
-      isNull(),
-    )
-
-    val actionedBy = if (eventAudit.actionedBy.userType == STAFF) eventAudit.actionedBy.userName else eventAudit.actionedBy.bookerReference
-    val eventsMap = mutableMapOf(
-      "reference" to cancelledVisit.reference,
-      "applicationReference" to cancelledVisit.applicationReference,
-      "prisonerId" to cancelledVisit.prisonerId,
-      "prisonId" to cancelledVisit.prisonCode,
-      "visitStatus" to cancelledVisit.visitStatus.name,
-      "visitSubStatus" to cancelledVisit.visitSubStatus.name,
-      "visitRestriction" to cancelledVisit.visitRestriction.name,
-      "visitStart" to cancelledVisit.startTimestamp.format(DateTimeFormatter.ISO_DATE_TIME),
-      "visitEnd" to cancelledVisit.endTimestamp.format(DateTimeFormatter.ISO_DATE_TIME),
-      "visitType" to cancelledVisit.visitType.name,
-      "visitRoom" to cancelledVisit.visitRoom,
-      "hasPhoneNumber" to ((cancelledVisit.visitContact.telephone != null).toString()),
-      "hasEmail" to ((cancelledVisit.visitContact.email != null).toString()),
-      "totalVisitors" to (cancelledVisit.visitors.size.toString()),
-      "visitors" to objectMapper.writeValueAsString(visitors),
-      "actionedBy" to actionedBy,
-      "source" to eventAudit.actionedBy.userType.name,
-      "applicationMethodType" to eventAudit.applicationMethodType.name,
-      "outcomeStatus" to cancelledVisit.outcomeStatus!!.name,
-    )
-
-    verify(telemetryClient, times(1)).trackEvent(type.eventName, eventsMap, null)
   }
 }
